+++ vue.config.js
@@ -78,6 +78,7 @@ module.exports = {
         '@gilt-page': `${root}/vue/pages/gilt`,
         '@society-page': `${root}/vue/pages/society`,
         '@democracy-page': `${root}/vue/pages/democracy`,
+        '@council-page': `${root}/vue/pages/council`,
       },
     },
   },